+++ OperatorsAndConditions/06.PointDistance/Startup.cpp
@@ -5,10 +5,10 @@ using namespace std;
 
 int main()
 {
-	int x1 = 0, y1 = 0, x2 = 0, y2 = 0, x = 0, y = 0;
+	double x1 = 0, y1 = 0, x2 = 0, y2 = 0, x = 0, y = 0;
 	cin >> x1 >> y1 >> x2 >> y2 >> x >> y;
 
-	int distance = 0;
+	double distance = 0;
 
 	if (x >= x1 && x <= x2 && y >= y1 && y <= y2)
 	{
